requirements.txt
Django>=4.2
channels>=4.0
daphne>=4.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
djangorestframework>=3.14
drf-yasg>=1.21
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
Pillow>=10.0
channels-redis>=4.0
celery>=5.3
whitenoise>=6.6
gunicorn>=21.2
